fix: added OPTIONS handling for CORS preflights (#17)
This version contains the inclusion of special headers, needed for CORS handling. These headers are required after an integration test with "pagoPA Shared Toolbox" portal, due to the fact that preflight request does not permits to Mocker response to be correctly handled.  
Also, a new header named X-Powered-By is added in order to add the possibility to better track mock responses.