fix: Resolved wrong MockerNotCompliantRequestException if request body is not passed (#20)
This commit contains a fix made in order to resolve a bug on mock condition evaluation found during an integration test. Previously, if a request does not contains a body and a mock condition that checks body field is defined, the same check ended prematurely because the Mocker throws a `MockerNotCompliantRequestException` exception.  
Following the best practice of default rules definition (the rule to be always executed if the request is not compliant to any previous rule), this intermediate exception is not required anymore because the check at worst returns a custom body. If no default rule is defined, the Mocker throws the `MockerNotCompliantRequestException` exception anyway at the end of all conditions check, so the unnecessary intermediate throw is removed.  
Another change made is related to the default content-type in request. During the request extraction made by proxy, if no "Content-Type" header is set, it will be assigned a default value (`text/plain`) in order to avoid unparseable body and consider them as plain text.

#### List of Changes
- Removed `MockerNotCompliantRequestException` throw if request body is not passed and there is a mock condition that make evaluation on its fields
- Force set default content-type header in request, if this header is not passed